+++ src/tsm/__init__.py
@@ -1,8 +1,11 @@
 """Time series modeling module."""
 
-from .prediction import GARCHPredictor, ARIMAGARCHPredictor
+from .prediction import ARIMAGARCHPredictor, GARCHPredictor, MarkovSwitchingGARCHPredictor
+from .regime_detector import RegimeDetector
 
 __all__ = [
     "GARCHPredictor",
     "ARIMAGARCHPredictor",
+    "MarkovSwitchingGARCHPredictor",
+    "RegimeDetector",
 ]